refactor(bartworks): isolate recipe rewriting
